Add embeds to edit log
